refactor : ChatWidget 증복되는 컴포넌트 분리했습니다
