+++ src/2_Widget/ChatWidget/ui/ChatBody/type.d.ts
@@ -1,0 +1,9 @@
+type ChatBodyProps = {
+  error: string | null;
+  chatLog: ChatMessageSocket[];
+  myNickname?: string | null;
+  myProfile?: string | null;
+  messagesEndRef: React.RefObject<HTMLDivElement>;
+  connectionStatus: string;
+  sendMessage: (data: ChatFormData) => void;
+};
